Revert "Add draft of clearchainctl"
This reverts commit 6b73f6b3786ca23744b5fa7b8be179f4ad7d35a3.
