--- commands/version.go
@@ -1,27 +1,0 @@
-package commands
-
-import (
-	"fmt"
-	"os"
-
-	"github.com/spf13/cobra"
-	"github.com/tendermint/clearchain"
-)
-
-var (
-	// VersionCmd prints the program's version to stderr and exits.
-	VersionCmd = &cobra.Command{
-		Use:   "version",
-		Short: "Print the app version",
-		Run:   doVersionCmd,
-	}
-)
-
-func doVersionCmd(cmd *cobra.Command, args []string) {
-	v := clearchain.Version
-	if len(v) == 0 {
-		fmt.Fprintln(os.Stderr, "unset")
-		return
-	}
-	fmt.Fprintln(os.Stderr, v)
-}
